requirements.txt
instabot==0.35.0
python-dotenv=0.10.1
